+++ src/events/render-advert.ts
@@ -187,6 +187,8 @@ const renderAdvert = (
 
 	return getAdIframe(advert.node)
 		.then((isRendered) => {
+			const creativeTemplateId =
+				slotRenderEndedEvent.creativeTemplateId ?? undefined;
 			const callSizeCallback = () => {
 				if (advert.size) {
 					/**
@@ -214,7 +216,7 @@ const renderAdvert = (
 					: Promise.resolve();
 
 			return callSizeCallback()
-				.then(() => renderAdvertLabel(advert.node))
+				.then(() => renderAdvertLabel(advert.node, creativeTemplateId))
 				.then(() => addContainerClass(advert.node, isRendered))
 				.then(addRenderedClass)
 				.then(() => isRendered);