+++ src/insert/spacefinder/spacefinder-debug-tools.ts
@@ -108,6 +108,8 @@ const addHoverListener = (
 					opponent.element,
 					`${opponent.actual}px/${opponent.required}px`,
 				);
+			} else {
+				addOverlay(opponent.element, 'Blocking element');
 			}
 		});
 
@@ -159,7 +161,10 @@ const annotateExclusions = (
 const annotateBody = (rules: SpacefinderRules, pass: SpacefinderPass) => {
 	const body = document.querySelector(rules.bodySelector) as HTMLElement;
 	body.id = 'sfdebug-body';
-	body.setAttribute(`data-sfdebug-min-above-${pass}`, `${rules.minAbove}px`);
+	body.setAttribute(
+		`data-sfdebug-min-above-${pass}`,
+		`${rules.minDistanceFromTop}px`,
+	);
 };
 
 const annotate = (