+++ src/lib/header-bidding/utils.ts
@@ -1,8 +1,6 @@
 import { isString } from '@guardian/libs';
 import { once } from 'lodash-es';
 import { createAdSize } from 'core/ad-sizes';
-import { isUserInVariant } from 'experiments/ab';
-import { prebidMagnite } from 'experiments/tests/prebid-magnite';
 import {
 	getCurrentTweakpoint,
 	matchesBreakpoints,
@@ -164,10 +162,6 @@ export const shouldIncludeTrustX = (): boolean => isInUsOrCa();
 export const shouldIncludeTripleLift = (): boolean =>
 	isInUsOrCa() || isInAuOrNz();
 
-export const shouldIncludeAdYouLike = (
-	slotSizes: HeaderBiddingSize[],
-): boolean => containsMpu(slotSizes);
-
 // TODO: Check is we want regional restrictions on where we load the ozoneBidAdapter
 export const shouldUseOzoneAdaptor = (): boolean =>
 	!isInCanada() &&
@@ -193,16 +187,10 @@ export const shouldIncludeImproveDigitalSkin = (): boolean =>
  */
 export const shouldIncludeCriteo = (): boolean => !isInAuOrNz();
 
-/**
- * Determine whether to include Smart as a prebid bidder
- */
-export const shouldIncludeSmart = (): boolean => isInUk() || isInRow();
-
 export const shouldIncludeKargo = (): boolean => isInUsa();
 
 export const shouldIncludeMagnite = (): boolean =>
-	!!window.guardian.config.switches.prebidMagnite &&
-	isUserInVariant(prebidMagnite, 'variant');
+	!!window.guardian.config.switches.prebidMagnite;
 
 export const shouldIncludeMobileSticky = once(
 	(): boolean =>